+++ components/Form/FormSelect/index.tsx
@@ -1,4 +1,4 @@
-import React, { useState, forwardRef } from 'react';
+import React, { forwardRef } from 'react';
 import type { ListBoxItemProps, SelectProps, ValidationResult } from 'react-aria-components';
 import { Button, FieldError, Label, ListBox, ListBoxItem, Popover, Select, SelectValue, Text } from 'react-aria-components';
 import styles from './formSelect.module.scss';